twitter-bot/requirements.txt
tweepy == 3.5.0
pyyaml == 3.10
